# AppKube-common-security-rbac-services/Api-Specifications/apispecs.yml
openapi: 3.0.3
info:
  title: Common security rbac services
  version: 1.0.0
  description: API for managing organizations, policies, and emails
servers:
  - url: http://localhost:8080/api/v1
...
paths:
  /organization/profile-by-id/{id}:
    get:
      tags:
        - Organization
      summary: Retrieve organization profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          description: Organization not found
        '417':
          description: Expectation failed
  /organization/profile-by-name/{name}:
    get:
      tags:
        - Organization
      summary: Retrieve organization profile by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          description: Organization not found
        '417':
          description: Expectation failed
  /organization/getAllOrganizations:
    get:
      tags:
        - Organization
      summary: Retrieve all organizations
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
        '404':
          description: Organizations not found
  /organization/id/{id}:
    get:
      tags:
        - Organization
      summary: Retrieve organization by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          description: Organization not found
  /organization/getOrganizationByUserName:
    get:
      tags:
        - Organization
      summary: Retrieve organization by user name
      parameters:
        - name: userName
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          description: Organization not found
  /organization/name/{organizationName}:
    get:
      tags:
        - Organization
      summary: Retrieve organization by name
      parameters:
        - name: organizationName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          description: Organization not found
  /organization/add:
    post:
      tags:
        - Organization
      summary: Add a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organization'
      parameters:
        - name: pushToCmdb
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '417':
          description: Expectation failed
  /organization/push-to-cmdb/{organizationName}:
    post:
      tags:
        - Organization
      summary: Push organization to CMDB
      parameters:
        - name: organizationName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          description: Organization not found
...
components:
  schemas:
    Organization:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        fax:
          type: string
        dateOfEstablishment:
          type: string
          format: date-time
        status:
          type: string
        cmdbOrgId:
          type: integer
          format: int64
        fileName:
          type: string
        fileStorageLocationType:
          type: string
        profileImage:
          type: string
...